fix: update Dockerfile and Makefile to use 'docker compose' syntax; refactor dependencies in pyproject.toml and uv.lock; add tests for plotting functions
