+++ Makefile
@@ -24,13 +24,13 @@ lint-unsafe-fix:
 pre-ci: format lint test
 
 compose-up-build:
-	docker-compose -f docker-compose.build.yml up -d --build
+	docker compose -f docker-compose.build.yml up -d --build
 
 compose-up:
-	docker-compose up -d
+	docker compose up -d
 
 compose-down:
-	docker-compose down
+	docker compose down
 
 compose-clean:
-	docker-compose down --volumes --remove-orphans
+	docker compose down --volumes --remove-orphans
