+++ pyproject.toml
@@ -7,18 +7,22 @@ requires-python = ">=3.13"
 dependencies = [
     "loguru>=0.7.3",
     "aiohttp>=3.11.14",
-    "pytest-asyncio>=0.25.3",
     "pandas>=2.2.3",
-    "pytest>=8.3.5",
     "python-dotenv>=1.0.1",
     "SQLAlchemy>=2.0.39",
-    "ruff>=0.11.2",
-    "aioresponses>=0.7.8",
     "pydantic>=2.10.6",
-    "pytest-mock>=3.14.0",
     "seaborn>=0.13.2",
     "matplotlib>=3.8.3",
     "pillow>=11.1.0",
     "flask>=3.0.3",
     "gunicorn>=22.0.0",
+]
+
+[dependency-groups]
+dev = [
+    "pytest>=8.3.5",
+    "pytest-asyncio>=0.25.3",
+    "pytest-mock>=3.14.0",
+    "ruff>=0.11.2",
+    "aioresponses>=0.7.8",
 ]
