+++ tests/visual/test_plotting.py
@@ -314,3 +314,43 @@ def test_plot_path_rarity_distribution():
         # Verify
         mock_catplot.assert_called_once()
         assert result is mock_fig
+
+
+@pytest.mark.visual
+def test_plot_path_balance_evolution():
+    """Test plotting path balance evolution."""
+    # Create test data with the correct columns
+    df = pd.DataFrame(
+        [
+            {"Version": "1.0", "Path": "Hunt", "count": 1},
+            {"Version": "1.0", "Path": "Harmony", "count": 1},
+            {"Version": "1.1", "Path": "Hunt", "count": 2},
+            {"Version": "1.1", "Path": "Harmony", "count": 2},
+        ]
+    )
+
+    # Patch the seaborn lineplot and pyplot
+    with patch("seaborn.lineplot") as mock_lineplot:
+        # Setup matplotlib mock
+        with (
+            patch("matplotlib.pyplot.subplots") as mock_subplots,
+            patch("matplotlib.pyplot.tight_layout") as mock_tight_layout,
+        ):
+            mock_fig = MagicMock()
+            mock_ax = MagicMock()
+            mock_subplots.return_value = (mock_fig, mock_ax)
+
+            # Execute the function
+            from hsrws.visual.plotting.plotting_line import plot_path_balance_evolution
+
+            result = plot_path_balance_evolution(df, "Patch (1.6)")
+
+            # Verify
+            mock_lineplot.assert_called_once()
+            # Verify legend placement
+            mock_ax.legend.assert_called_once_with(
+                title="Path", bbox_to_anchor=(1.05, 1), loc="upper left"
+            )
+            # Verify tight_layout is called
+            mock_tight_layout.assert_called_once()
+            assert result is mock_fig
